dev_requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pluggy==0.6.0
py==1.5.2
PyHamcrest==1.9.0
pytest==3.4.0
requests==2.18.4
requests-mock==1.4.0
six==1.11.0
urllib3==1.22
